pyproject.toml
[project]
name = "embedding-pca-during-training"
version = "0.1.0"
description = "PCA measurements of embedding layers across training checkpoints for Pythia and OLMo model families."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scikit-learn",
  "safetensors",
  "huggingface_hub",
  "pyyaml",
  "tqdm",
]

[project.scripts]
run-model = "src.scripts.run_model:main"
run-family = "src.scripts.run_family:main"
summarize-results = "src.scripts.summarize_results:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
